pyproject.toml
[project]
name = "jmake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "notion-client>=2.4.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
